meu curso NodeJs

+++ 2_FUNDAMENTOS/6_pratica_com_args/index.js
@@ -6,9 +6,9 @@ const minimist = require("minimist")
 
 const soma = require ('./soma.js').soma
 const args = minimist(process.argv.slice(2))
-
-const a = parseInt(args['a'])
-const b = parseInt(args['b'])
+ 
+const a = parseFloat(args['a'])
+const b = parseFloat(args['b'])
 
 soma (a,b)
